+++ client/src/features/add-bet/add-bet-inputs/TypeInput.tsx
@@ -2,9 +2,9 @@ import "./TypeInput.css";
 
 import { ChangeEvent } from "react";
 
-import { Select } from "../../../../components/ui/inputs/Select";
-import { LimitedBetType } from "../../../../utils/enums";
-import { BetDetails } from "../../../../utils/types";
+import { Select } from "@/components/ui/inputs/Select";
+import { LimitedBetType } from "@/utils/enums";
+import { BetDetails } from "@utils/types";
 
 type TypeInputProps = {
   handleSelectChange: (e: ChangeEvent<HTMLSelectElement>) => void;
