+++ client/src/features/add-bet/add-bet-inputs/tests/SelectionInput.spec.tsx
@@ -2,9 +2,9 @@ import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 
-import { initialBetValues } from "../../../../../utils/defaults/defaults";
-import { inputErrors } from "../../../../../utils/defaults/errors";
-import { AddBetForm } from "../../AddBetForm";
+import { initialBetValues } from "@/utils/defaults/defaults";
+import { inputErrors } from "@/utils/defaults/errors";
+import { AddBetForm } from "../../forms/add-bet-form/AddBetForm";
 
 const user = userEvent.setup();
 const mockFn = vi.fn();
@@ -17,7 +17,7 @@ beforeEach(() => {
       modifyIndex={null}
       setModifyIndex={mockFn}
       disabled={false}
-    />,
+    />
   );
 });
