+++ client/src/features/add-bet/forms/add-bet-form/BetSelection.tsx
@@ -5,8 +5,8 @@ import { Dispatch } from "react";
 import { faXmark } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 
-import { BetDetails } from "../../../utils/types";
-import { isBetBuilderType } from "../betUtils";
+import { BetDetails } from "../../../../utils/types";
+import { isBetBuilderType } from "../../../../pages/add-bet/betUtils";
 
 interface BetSelectionProps {
   details: BetDetails;
