+++ client/src/pages/add-bet/index.ts
@@ -1,11 +1,4 @@
-export { DateInput } from "./add-bet-form/add-bet-inputs/DateInput";
-export { FreeLiveInput } from "./add-bet-form/add-bet-inputs/FreeLiveInput";
-export { MatchInput } from "./add-bet-form/add-bet-inputs/MatchInput";
-export { OddsInput } from "./add-bet-form/add-bet-inputs/OddsInput";
-export { SelectionInput } from "./add-bet-form/add-bet-inputs/SelectionInput";
-export { TypeInput } from "./add-bet-form/add-bet-inputs/TypeInput";
-export { BetbuilderInput } from "./add-bet-form/add-bet-inputs/BetbuilderInput";
-export { AddBetForm } from "./add-bet-form/AddBetForm";
+export { AddBetForm } from "../../features/add-bet/forms/add-bet-form/AddBetForm";
 
 export { FinishBetForm } from "./finish-bet-form/FinishBetForm";
 export { BookmakerInput } from "./finish-bet-form/finish-bet-inputs/BookmakerInput";
@@ -14,4 +7,3 @@ export { NotesInput } from "./finish-bet-form/finish-bet-inputs/NotesInput";
 export { SportInput } from "./finish-bet-form/finish-bet-inputs/SportInput";
 
 export { FinishBetButtons } from "./FinishBetButtons";
-// export { AddBet } from "./AddBet";
